Mock the virtual flow behavior before working on the real thing
Added testfx, scenicview and log4j dependencies
All changes made to Cell, SimpleCell, VirtualFlow and VirtualFlowContainer are mostly for testing purpose or experimental

MockVirtualFlow:
This is a "simulation" of how a virtual flow should behave. A Weight object is like a Cell, it has a name, a value (which represents the height), and a boolean property for debug purposes.
We define heights of 150 for the virtual flow and 300 for the weights container (10 weights with a fixed height of 30).
To mimic the scrollbar and JavaFX bounds, we add a key event filter for arrows up and down that changes the containerBounds property accordingly, by the value specified by scrollUnit (10 by default)
The visibleIndexes property keeps track of the range on indexes that are visible, but it doesn't take into account the overscan.
The overscan is the number of extra indexes to add, cells that won't be visible but laid out anyway, ready to be shown on scroll.

This mock system makes it easier to understand how the real flow should work because we don't work with JavaFX nodes, it's just the concept
Added tests for the mock system (most of them fail as of now but the system is still buggy and maybe overcomplicated. Pushing these changes to work on another machine)