Completed the mock system with an animated virtual flow
Got it. I finally have a working virtual flow, you can check its behavior by running AnimatedMockVirtualFlowApp

Next step will be to work on the real actual thing (so, cell factories, concept of Cell behavior and properties, layout, scroll bars and so on)

Ditched the old mock system and all its tests in favor of AnimatedMockVirtualFlow, all tests to check if the overscan works as expected are passing.